+++ src/cs2fixes.h
@@ -92,7 +92,7 @@ public: // MetaMod API
 	std::uint64_t GetAdminFlags(std::uint64_t iSteam64ID) const override;
 	bool SetAdminFlags(std::uint64_t iSteam64ID, std::uint64_t iFlags) override;
 	int GetAdminImmunity(std::uint64_t iSteam64ID) const override;
-	bool SetAdminImmunity(std::uint64_t iSteam64ID, std::uint16_t iImmunity) override;
+	bool SetAdminImmunity(std::uint64_t iSteam64ID, std::uint32_t iImmunity) override;
 
 public:
 	const char* GetAuthor();
